Stoned again... and again...

+++ src/main/java/de/spring/elolink_spring/service/ChatService.java
@@ -3,15 +3,15 @@ package de.spring.elolink_spring.service;
 import de.spring.elolink_spring.dtos.ChatDto;
 import de.spring.elolink_spring.entity.Chat;
 import de.spring.elolink_spring.repository.ChatRepository;
-import io.netty.handler.codec.dns.DnsResponseCode;
+import de.spring.elolink_spring.repository.UserRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import reactor.core.publisher.Mono;
-import reactor.core.scheduler.Scheduler;
-import reactor.core.scheduler.Schedulers;
 
-import java.net.http.HttpResponse;
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.List;
@@ -23,15 +23,22 @@ public class ChatService {
     @Autowired
     private ChatRepository chatRepository;
 
-    //Todo: Verification!!!!!
+    @Autowired
+    private UserRepository userRepository;
 
     @Transactional
-    public String addChat(ChatDto chatDto) {
+    public ResponseEntity<String> addChat(ChatDto chatDto, Authentication authentication) {
+        if (!userRepository.existsByUserName(chatDto.getSender())){
+            return new ResponseEntity<>("User doesn't exist", HttpStatus.BAD_REQUEST);
+        }
+        if (!Objects.equals(chatDto.getSender(), authentication.getName())){
+            return new ResponseEntity<>("Requested username and authenticated username don't match!", HttpStatus.UNAUTHORIZED);
+        }
         Chat chat = Chat.fromDto(chatDto);
         chat.setId((long) (null == chatRepository.findMaxId() ? 0 : chatRepository.findMaxId() + 1));
         chatRepository.save(chat);
-        System.out.println("#Added Chat '" + chatDto.getMessage() + "' from Sender " + chatDto.getSender());
-        return "200";
+        System.out.println("#Added chat '" + chatDto.getMessage() + "' from sender " + chatDto.getSender());
+        return new ResponseEntity<>("Added chat from user " + chat.getSender(), HttpStatus.OK);
     }
 
     public Mono<List<ChatDto>> getConversation(String uuid1, String uuid2) {
@@ -51,17 +58,29 @@ public class ChatService {
     }
 
     @Transactional
-    public String deleteChat(String sender, String receiver, String timestamp){
+    public ResponseEntity<String> deleteChat(String sender, String receiver, String timestamp, Authentication authentication){
+        if (!userRepository.existsByUserName(sender)){
+            return new ResponseEntity<>("User doesn't exist", HttpStatus.BAD_REQUEST);
+        }
+        if (!Objects.equals(sender, authentication.getName())){
+            return new ResponseEntity<>("Requested username and authenticated username don't match!", HttpStatus.UNAUTHORIZED);
+        }
         chatRepository.deleteBySenderAndReceiverAndTimestamp(
                 sender,
                 receiver,
                 timestamp);
         System.out.println("Deleted Chat from '"  + sender + "' to '" + receiver + "' from " + timestamp);
-        return "200";
+        return new ResponseEntity<>("Deleted chat from user " + sender, HttpStatus.OK);
     }
 
     @Transactional
-    public String updateChat(String sender, String receiver, String timestamp, ChatDto chatDto){
+    public ResponseEntity<String> updateChat(String sender, String receiver, String timestamp, ChatDto chatDto, Authentication authentication){
+        if (!userRepository.existsByUserName(sender)){
+            return new ResponseEntity<>("User doesn't exist", HttpStatus.BAD_REQUEST);
+        }
+        if (!Objects.equals(sender, authentication.getName())){
+            return new ResponseEntity<>("Requested username and authenticated username don't match!", HttpStatus.UNAUTHORIZED);
+        }
         List<Chat> chats = chatRepository.findBySenderAndReceiverAndTimestamp(sender, receiver, timestamp);
         chats.forEach(chat -> {
             chat.setSender(chatDto.getSender());
@@ -74,9 +93,7 @@ public class ChatService {
             chatRepository.save(chat);
         });
         System.out.println("Updated Chat from '"  + sender + "' to '" + receiver + "' from " + timestamp);
-        return "200";
+        return new ResponseEntity<>("Updated chat from user " + sender, HttpStatus.OK);
     }
-
-
 
 }
